refactor: 검색 영화 처리 로직을 별도의 이벤트 핸들러로 분리
